+++ NEWS.md
@@ -1,4 +1,6 @@
 # isopleuros 1.0.0.9000
+## Internals
+* Use **tinytest** and **tinysnapshot** instead of **testthat** and **vdiffr**.
 
 # isopleuros 1.0.0